requirements.txt
Django==1.8.2
PyYAML==3.11
django-rest-swagger==0.3.1
djangorestframework==3.1.2
psycopg2==2.6
requests==2.7.0
ipython
django-extensions
